% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datiPorti-package.R
\docType{package}
\name{datiPorti-package}
\alias{datiPorti}
\alias{datiPorti-package}
\title{datiPorti: What the Package Does (One Line, Title Case)}
\description{
Dati per il progetto PORTI.
}
\author{
\strong{Maintainer}: First Last \email{raffaele.morelli@isprambiente.it} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
