% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_nhr.R
\docType{data}
\name{df_nhr}
\alias{df_nhr}
\title{Data of nominal brown shrimp prices by size}
\format{
A tibble with 243 rows and 10 variables:
\describe{
\item{pltnt}{chr l'inquinante}
\item{irec}{dbl }
\item{gt}{dbl }
\item{terminal}{dbl }
\item{Ship_categ}{dbl }
\item{ntug}{dbl }
\item{Emiss_HOT_MDGO}{dbl }
\item{Emiss_MAN_MDGO}{dbl }
\item{Emiss_tug_MDGO}{dbl }
\item{Emiss_Cter_MDGO}{dbl }
\item{Emiss_Cext_MDGO}{dbl }
\item{Emiss_HOT_BFO}{dbl }
\item{Emiss_MAN_BFO}{dbl }
\item{Emiss_tug_BFO}{dbl }
\item{Emiss_Cter_BFO}{dbl }
\item{Emiss_Cext_BFO}{dbl }
}
}
\source{
\url{https://www.pnas.org/content/114/7/1512/tab-figures-data}
}
\usage{
df_nhr
}
\description{
Shrimp are classified by size, 0-15 shrimp per pound, 15-20 shrimp per pound, etc. A smaller number per pound indicates larger shrimp. Nominal prices are total monthly value of brown shrimp andings within size class divided by total monthly landings within the size class.
}
\keyword{datasets}
